Update option args tests
